Implement common block object element accessor
