+++ ydb/library/yql/minikql/comp_nodes/mkql_block_tuple.h
@@ -5,7 +5,6 @@ namespace NKikimr {
 namespace NMiniKQL {
 
 IComputationNode* WrapBlockAsTuple(TCallable& callable, const TComputationNodeFactoryContext& ctx);
-IComputationNode* WrapBlockNth(TCallable& callable, const TComputationNodeFactoryContext& ctx);
 
 }
 }
