+++ ydb/library/yql/minikql/comp_nodes/ya.make.inc
@@ -16,6 +16,7 @@ SET(ORIG_SOURCES
     mkql_block_agg_sum.cpp
     mkql_block_coalesce.cpp
     mkql_block_exists.cpp
+    mkql_block_getelem.cpp
     mkql_block_if.cpp
     mkql_block_just.cpp
     mkql_block_logical.cpp
